Remove unnecessary workspace
This is noisy for Carthage and Xcode. Since the projects have the schemes those are all the same name, Xcode can't pick up the correct project to build with for the command: `xcodebuild -workspace AudioKit.xcworkspace -scheme AudioKit`. Removing the workspace, Carthage would use a command such as: `xcodebuild -project "AudioKit For OSX.xcodeproj" -scheme AudioKit`.