I've refined the fill buffer route more now. Also, I've changed the name of setPCMBuffer() to updatePCMBuffer(), which is more accurate to what it's doing.
The buffer is filled only on these occasions now:
* initialize()
* set startTime
* set endTime
* on immediate playback within a sound file (when the startTime is being set on the fly)

Previously the buffer was being filled with the same data over and over (on stop() which was called too often) even if it didn't need to change. So, this should be more efficient. I've tested with 20 concurrent samples all playing at the same time and it was solid. I do think the current implementation is better suited towards shorter sound files as Laurent mentioned. More on that soon...