Sporth Custom Ugen: Renamed fp to funcPtr to remove ambiguity.
This is a minor nitpick. "fp", presumably standing for "function pointer", is a
variable name traditionally used for file i/o operations as the "file pointer".
To avoid the confusion, and since it's not obvious that plumber_dyn_func is a
function pointer, I have renamed fp to funcPtr. It is clearer, and adheres to
the coding style that surrounds it.