Refactored for offline render compatibility, and for timelier synchronization of files. Separated a new function schedule(from time: Double, to endTime: Double, avTime: AVAudioTime? ) from existing play(from time: Double, to endTime: Double, avTime: AVAudioTime? ) . Added play(at when: AVAudioTime?). start() now calls play(at: nil) internally.
AVAudioPlayer.play() will block until the next render cycle completes.  When performing an offline render, the rendering is paused and you must provide an AVAudioTime with a valid sampleTime or the player will block indefinitely.  AKAudioPlayer.play(at:) facilitates this need.
The combination of schedule(from:...) and play(at:) allows for files to be prepared before setting the avTime that they'll be synchronized with.  Once files are prepared, you can start them one (or two to be safe) render cycles in the future and they'll definitely be ready.