Safeguard for making connections to AVAudioMixer.
AVAudioMixer will crash if engine is started and connection is made to a
bus exceeding mixer's numberOfInputs. The crash only happens when
using the AVAudioEngine function that connects a node to an array of
AVAudioConnectionPoints and the mixer is one of those points. When
AVAudioEngine uses a different function that connects a node's output
to a single AVAudioMixerNode, the mixer's inputs are incremented to
accommodate the new connection. So the workaround is to create a
dummy node, make a connection to the mixer using the function that
makes the mixer create new inputs, then remove the dummy node so
that there is an available bus to connect to.