Built in a workaround for empty mixer on running engine bug.
If an AVAudioMixerNode's output connection is made while engine is running, and
there are no input connections on the mixer, subsequent connections made to
the mixer will silently fail.  A workaround is to connect a dummy node to the mixer
prior to making a connection, then removing the dummy node after the connection
has been made.