AudioKitCore::Sampler fix bug in note-retrigger
My anti-click efforts didn't quite work, because I forgot a return statement.
Now that this is working, I was able to simplify the logic for ADSREnvelope::start(). (What was there before was an earlier effort to resolve the click-on-retrigger problem.)
Fingers crossed, I think I've finally fixed it this time.