AKSampler: breaking change to raise efficiency
The per-voice filters in AKSampler weren't efficient due to frequent calls to math functions pow(), sin() and cos(). I eliminated the need for pow() with a breaking change to the API, where the filter resonance (traditionally expressed in dB) must be converted to a linear equivalent value. I eliminated the need for sin() and cos() by using a static sine table.