AKSampler: ADSR generator always does A, D, R phases
The question arose: how should an ADSR generator respond to a note-on immediately followed by note-off, i.e., an impulse? The original code would result in a note so quiet you couldn't hear it, because the ADSR would go straight to the Release phase after not having enough time in the Attack phase to build up any volume. This change makes the A and D phases happen regardless.