AKSampler: add "drum mode"
By default, note-on followed immediately by note-off would result in an inaudible note, because the amp envelope has not yet had a chance to rise from the initial level of 0, so it goes into the Release phase with whatever near-zero volume it has. When drum mode is enabled, the envelopes are forced to go entirely through their Attack and Decay phases before going to Release, even for arbitrarily short note durations.