Update FFTTap.swift
When we are not normalizing the fftData, we should still be returning values that result in strictly negative amplitudes. After looking into Apple's documentation, it seems that a scaling factor is required. I used 1/n (n = frameCount) for this factor and after performing my amplitude calculations in FFTView they matched the same amplitude values what I was getting from Ableton Live's spectrum analyzer.