Part 2 - API Cleaning
I tried to make the AudioKit API more clear for the user, so they don't need to think about whether or not there will be a split connection. This only happens when a node has 2 outputs and remains with 1 split connection after it is disconnected (which makes AVAudioEngineGraph.mm throw the false conditional error).

Please let me know if you have any other suggestions for a cleaner API, and I will happily fix it. This is all fun learning for me.