Remove assert and replace with conditional check, add channelCapabilities
The Plug-In Manager in Logic Pro X doesn't seem to like the assert statement. When auval throws a parameter address at the DSPBase to test the AudioUnitScheduleParameter, it hits the assert immediately, invalidating the plugin.

The auval tool will also fail when it tries to test running DSPBase with one channel. The user can now explicitly state the supportedLeft/supportedRight channel counts(which defaults to 2).