Do not run tap on main queue
This code is taking locks and doing calculations.
It is called many times per second and it shouldn't
run on the main queue.

If callsite needs to run it on the main queue, it should
be easy enough to add it on top of the tap. On another hand,
if we run it on main queue internally, callsite cannot decide to
run it on background queue.