Implement incremental connection to Mixer
Existing addInput strategy traverses existing connections and
reastablishes them.

This might not be always desirable.

The specific problem that I had was that I was adding node while having
parameter automation on another node that was already in the mixer.

Since we reset mixer volume while doing connections, parameter
automation was not working as expected.