Implement AUNewPitch audio unit node
Apple exposes two audio units behind kAudioUnitSubType_NewTimePitch.

One is AUNewTimePitch which is available as AVAudioUnitTimePitch and is
of type kAudioUnitType_FormatConverter.

Another one is AUNewPitch and this one is not available as a
AVAudioNode. The type of this unit is kAudioUnitType_Effect.

AUNewPitch can only perform pitch shifting and not time stretching.

Why you might want this:

- Simple effects can be wired directly to input node where time effects
  can't.
- Another reason is this bug: http://openradar.appspot.com/radar?id=5525654997041152