Fix sample rate in tests to 44100 (#2942)
* Check errors after AudioUnit interactions

I spent quite a bit of time debugging an issue where
`AudioUnitSetProperty` was returning error `OSStatus`.

Ideally, all of these would be marked as throws, so that call-site can
react appropriatelly to these errors. But I thought this is a good first
step to improve the current situation.

* Fix sample rate in tests to 44100

A bunch of tests were failing on my computer as `Settings.sampleRate`
changed and the result was different then the reference.