Allow control over v2 parameter invocation
When setting a parameter, AudioUnitSetParamter is inovked. This code was
added as a workaround for:
https://github.com/AudioKit/AudioKit/issues/2528, but generates a
warning for v3 AVAudioUnit's (the ones in AudioKitEX).

This PR allows parameter definition to decide wheater this workaround
needs to be applied.

Note: This PR keeps backwards compatibility, but I was not able to
reproduce this issue anymore even if removing the code completely (the
test passes), so Apple has likely fixed the issue.