Simplify code by removing unnecessary checks (#2964)
* Check errors after AudioUnit interactions

I spent quite a bit of time debugging an issue where
`AudioUnitSetProperty` was returning error `OSStatus`.

Ideally, all of these would be marked as throws, so that call-site can
react appropriatelly to these errors. But I thought this is a good first
step to improve the current situation.

* Simplify code by removing unnecessary checks

- Use `guard let` for less indented code
- Pass `AVAudioEngine` instance that we know we have to avoid
  unnecessary checks

* Switch to iPhone 16

iPhone 14 is not available anymore in CI

* Pass engine to disconnect

* Skip failing test