Respect outputFormat sample rate (#2970)
* Check errors after AudioUnit interactions

I spent quite a bit of time debugging an issue where
`AudioUnitSetProperty` was returning error `OSStatus`.

Ideally, all of these would be marked as throws, so that call-site can
react appropriatelly to these errors. But I thought this is a good first
step to improve the current situation.

* Respect outputFormat sample rate

When rendering in offline mode, respect outputFormat.sampleRate. This
will most commonly be Settings.sampleRate, but not necessarily.