Allow control over v2 parameter invocation (#2957)
* Check errors after AudioUnit interactions

I spent quite a bit of time debugging an issue where
`AudioUnitSetProperty` was returning error `OSStatus`.

Ideally, all of these would be marked as throws, so that call-site can
react appropriatelly to these errors. But I thought this is a good first
step to improve the current situation.

* Allow control over v2 parameter invocation

When setting a parameter, AudioUnitSetParamter is inovked. This code was
added as a workaround for:
https://github.com/AudioKit/AudioKit/issues/2528, but generates a
warning for v3 AVAudioUnit's (the ones in AudioKitEX).

This PR allows parameter definition to decide wheater this workaround
needs to be applied.

Note: This PR keeps backwards compatibility, but I was not able to
reproduce this issue anymore even if removing the code completely (the
test passes), so Apple has likely fixed the issue.

* Update to macOS 15

For some reason, macos-latest was using macOS 14

* Update CI device to latest

---------

Co-authored-by: Aurelius Prochazka <aure@users.noreply.github.com>